Mixed dtype: arithmetics

+++ chainerx_cc/chainerx/routines/math_test.cc
@@ -869,14 +869,14 @@ TEST_P(MathTest, DivideScalar) {
 TEST_P(MathTest, DivideInteger) {
     Array a = testing::BuildArray({3, 1}).WithData<int64_t>({1, 2, 3});
     Array b = testing::BuildArray({3, 1}).WithData<int64_t>({2, 2, 12});
-    Array e = testing::BuildArray({3, 1}).WithData<double>({0.5f, 1.f, 0.25f});
+    Array e = testing::BuildArray({3, 1}).WithData<float>({0.5f, 1.f, 0.25f});
 
     testing::CheckForward([](const std::vector<Array>& xs) { return std::vector<Array>{Divide(xs[0], xs[1])}; }, {a, b}, {e});
 }
 
 TEST_P(MathTest, DivideScalarInteger) {
     Array a = testing::BuildArray({3, 1}).WithData<int64_t>({1, 2, 3});
-    Array e = testing::BuildArray({3, 1}).WithData<double>({0.5f, 1.f, 1.5f});
+    Array e = testing::BuildArray({3, 1}).WithData<float>({0.5f, 1.f, 1.5f});
 
     testing::CheckForward([](const std::vector<Array>& xs) { return std::vector<Array>{Divide(xs[0], Scalar{2})}; }, {a}, {e});
 }
